Make MaybeWipeout function somewhat more efficient.
We only need to check the exclude prefs for bad extensions, not all extensions. Also, we don't need to check if the sideload wipeout has already happened each time we evaluate extensions (we can do that once and for all before we start evaluating extensions).

In other words, make an early return if the extension is of the wrong type/location before checking the prefs.

BUG=163697, 154624
TEST=None
Review URL: https://codereview.chromium.org/11439019