cc: Pass LayerTreeHostImpl to LayerImpl constructor
In order to allow for LayerImpl to notify its host when one of its property
changes, it needs to more robustly have its host set (which is not the case in
most tests).  To avoid a recursive pass when adding children or after layer
synchronization, instead just pass it to the constructor of LayerImpl, since a
given LayerImpl will never move from one host to another.

NOTRY=true
BUG=none

Review URL: https://chromiumcodereview.appspot.com/11472021