[cros] Deferred animated resize of multiline user pod row.
Pods are moved and resized only when mouse cursor leaves the entire pod row.
This makes removing a bunch of users easy as the cursor stays on the remove button (of the next pod) after removing a user.
Podrow and pods resize is animated.


BUG=122346


Review URL: https://chromiumcodereview.appspot.com/11478023