[Android] Mark final input events for current vsync period
Android delivers input events to applications right before their views
are redrawn. This is done to maximize the time an application has to
process input events and draw a new frame before the next vsync period
starts.

This patch conveys this information to the renderer so that it can use
it for more efficient render scheduling. Only
TouchMove/ScrollUpdate/PinchUpdate are annotated, because other events
may be synthesized by the gesture recognizer and sent in bursts.


Review URL: https://chromiumcodereview.appspot.com/11416320