Use MRU cache for ppapi local time zone offset queries.
The PPB_Flash_Proxy::GetLocalTimeZoneOffset() calls often arrive in a batch of
multiple calls at different time offsets, thus it is necessary to use an MRU
cache to hold multiple cached local time zone offsets.

My experiments show that a 6-deep MRU cache is adequate to fix the performance
problems with this particular issue, but making it 100 entries just to be safe.
The memory footprint is negligible.

BUG=160203
TEST=Ameba Pigg gameplay


Review URL: https://chromiumcodereview.appspot.com/11442021