NaCl: Move buildbot_chrome_nacl_stage.py into the Chromium repo
This script controls what browser tests we run (via Scons, in the
nacl_integration Buildbot step), so it really belongs in the Chromium
repo.

I have made small changes to the script to work in the new location,
and to fix line-length presubmit warnings.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2957
TEST=nacl_integration

Review URL: https://codereview.chromium.org/11475034