Rename discardFramebufferEXT to discardBackbufferCHROMIUM to avoid collision with real discardFramebufferEXT
Calling the (real) discardFramebufferEXT in the compositor is a performance win, at least on tilers. However
currently it can't be called because we alias the name for a different API that controls backbuffer
allocation, which I don't want to change.  This renames the code to make backbuffer control a completely
separate API so I can pipe GL_EXT_discard_framebuffer in fully as a follow-up.

Depends on https://bugs.webkit.org/show_bug.cgi?id=104316

BUG=


Review URL: https://chromiumcodereview.appspot.com/11474014