NaCl: Extend nacl_integration to test the old SRPC-based PPAPI proxy
Currently nacl_integration is only testing the new Chrome-IPC-based
PPAPI proxy.  But until we are sure that it is OK to deploy the new
proxy in stable builds, we also want to be testing the old proxy
side-by-side.

I have used an environment variable to plumb the
"--enable-nacl-srpc-proxy" argument through because adding an argument
to Scons currently requires changing ACCEPTABLE_ARGUMENTS on the NaCl
side.

BUG=116317
TEST=nacl_integration


Review URL: https://chromiumcodereview.appspot.com/11469041