Move ExtensionSystemImpl::extension_warning_service_ to ExtensionSystemImpl::shared
Move ExtensionSystemImpl::extension_warning_service_ to ExtensionSystemImpl::shared to prevent a crasher in Guest Mode of ChromeOS. The Guest Mode was a OTR profile that did not have a initialized ExtensionWarningService but was nevertheless used to generate the chrome://extensions page. By moving the EWS into a data container that is shared between the regular profile and OTR profiles (such as the Guest Profile) we fix this problem.


BUG=163007


Review URL: https://chromiumcodereview.appspot.com/11472020