cc: Fix impl-side painting rasterization at >1 contents scales
Previously there was a content space vs. layer space mixup.  cc::Tile had a
member variable that was labeled rect_in_layer_space, but the tilings were
making a content space rect.  It's actually impossible to have an int rect in
layer space for most content spaces, so this got renamed to just be
content_space rect.

The actual bug was that this rect was being used to check if a given picture's
layer rect intersected it, so as soon as you got over bounds() / content_scale
away from the origin, the tile was getting nothing rastered to it and whatever
memory was in the canvas was being used, which happened to be old tiles.

R=aelias@chromium.org
BUG=164750


Review URL: https://chromiumcodereview.appspot.com/11466032