Add observer interface to PrerenderContents.
This CL adds a simple Observer interface to PrerenderContents, and makes the PrerenderTracker an observer. The PrerenderHandle will become an observer in a soon to be uploaded CL....

The most interesting change here is that the prerender_unittest is now testing with a tracker, so it has to pass in reasonable route IDs.

R=mmenke@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/11348357