output_configurator: Skip reallocating the framebuffer when not needed.
When configuring outputs, don't reallocate the framebuffer if it's already the
right size. This avoids the ConfigureCrtc calls which turn off all the
displays. This reconfiguration can take up to two seconds on Link, because
each panel on/off takes 500ms there.

This speeds up modesetting, and suspend/resume times. Note that this applies
to all Chrome OS machines, so will result in improving times across the board,
but the panel timings aren't as bad on other platforms so it makes less of a
difference.

BUG=chrome-os-partner:13364
TEST=by hand

Change-Id: Ia8344aa689c0307e8a1c85a5a2b75af6fb4b8371


Review URL: https://chromiumcodereview.appspot.com/11471039