bookmarks: Make BookmarkModel not send notifications through content::NotificationService.
The lack of type safety and the endless dispatching by
notification ID makes it a pretty bad interface.

So NotificationService is considered pretty evil and should be used in less modules, not more.

This removes the chrome::NOTIFICATION_URLS_STARRED notification from BookmarkModel and make the clients
listen to BookmarkModel changes through BaseBookmarkModelObserver which is clearer.

BUG=144783
R=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11450002