Cache the system version, numbers, and name, as the OS does not.
During startup, the system name and/or version is requested a number of times, and each call is 1ms or more.  Changed the numbers parsing to use std::string instead of NSString so that the static variable could be an std::string, avoiding having to make the conversion every function call.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/11437046