Re-land fix for Media Foundation crash.  Remove Vista support. Original description:
Do a thorough check for the availability of the Media Foundation DLLs.
Comments from code:
On Vista this API is an optional download but the API is
advertised as a part of Windows 7 and onwards. However,
we've seen that the required DLLs are not available in some
Win7 distributions such as Windows 7 N and Windows 7 KN.

BUG=164654
TEST=Run media_unittests. See more details in the bug report.

TBR=mflodman


Review URL: https://chromiumcodereview.appspot.com/11500009