Double the buildHorizon on the Try Server.
At 6000 builds, it's roughly a few days worth. Note that logs are not kept,
only the build objects, which are significantly smaller. Each build file is
~10kb to ~100kb so we can expect an increase of disk usage around 10gb

Otherwise, the CQ queries try jobs that are pruned already, causing the CQ to
retry them. It's potentially unnecessary retries because the CQ can't tell if
the try job itself would not be expired.

R=cmp@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/11308242