Fix tooltips in ash.
ash::internal::TooltipController was conditionally setting InitParams::parent,
which was putting tooltips in the wrong part of the ash hierarchy on some
platforms, while on others, there was no contextual information. Pass
InitParams::context on all platforms instead; this is what context is for.

BUG=151718, 161882


Review URL: https://chromiumcodereview.appspot.com/11516007