Extend virtual context MakeCurrent fastpath to offscreen surfaces.
Similar to transport surfaces with an explicit fbo backing,
we can avoid the call into the driver to make the surface current
since the command decoder guarantess an fbo backing for offscreen
surfaces.

On Android, this specifically affects canvas contexts and the helper
context used on the UI thread for mailbox texture exchange.

BUG=160710


Review URL: https://chromiumcodereview.appspot.com/11537014