Callback to function InitializeRegisteredDeviceWallpaper after cros settings can be trusted
InitializeWallpaper is being called in a very early stage(shell initialization). At such point, the cros settings may not be validated, so the value maybe wrong. Function InitializeWallpaper, however, depends on kAccountsPrefShowUserNamesOnSignIn to behavior correctly.

This CL made a callback to InitializeWallpaper if we detect that the settings is not yet validated.

BUG=162760

Review URL: https://codereview.chromium.org/11299304