Improve K-mean dominant color selection for favicons
After running the favicon through K-mean clustering, find a color from the
image that is closest to the cluster center. This usually makes the colors
less "muddy", as K-mean clusters converge on an average color. It improves
selection for sites like reddit and imgur.

BUG=162315
TEST=added ui_unittests ColorAnalysisTest, FindClosestColor


Review URL: https://chromiumcodereview.appspot.com/11485011