Changes to support building a libchromeview.so with prebuilt sources.
Change the JNI_OnLoad function such that it's always included the final
shared_library artifact. This is necessary because of the use of
exclude-libs=ALL which otherwise would hide that symbol. Instead, it
relies on a static create function for ChromeMainDelegateAndroid which
can be provided with other code in a prebuilt static library.


Review URL: https://chromiumcodereview.appspot.com/11419317