Properly disable accelerated compositing for extension background pages. Since these pages are not expected to ever render, there's no reason for the additional overhead.
In extension_webkit_preferences::SetPreferences() we were already trying to disable force_compositing_mode
but that wasn't taking effect as it was being reset later on in ChromeContentBrowserClient::OverrideWebkitPrefs.
I'm not sure why we would want to allow for any kind of GPU acceleration on background tabs so this patch
disables it all.

BUG=163698


Review URL: https://chromiumcodereview.appspot.com/11299288