Enable virtual context on IMG devices only.
This adds booleans to feature_info for all the common Android vendors
and turns on virtual contexts for IMG only, as there are remaining bugs
with the other architectures (and IMG is the main reason we are working
on virtual contexts to begin with).  Setting the command-line flag will
force-enable virtual contexts on any architecture.

BUG=160710,165531

Review URL: https://chromiumcodereview.appspot.com/11555010