Ignore plugin MIME entries containing WebPluginTypeEnabled=0 on Mac
Mac plugin plist declarations can include a WebPluginTypeEnabled key
which can be set to 0 to disable the plugin handling that MIME type.
This key had not been honored before.

If the key is not present, the MIME type is assumed to be enabled.

BUG=144655
TEST=On chrome://plugins, QuickTime should not list anything whose
entry in the output of 'defaults read com.apple.quicktime.plugin.preferences'
has WebPluginTypeEnabled=0. Plugins without WebPluginTypeEnabled
in their plist entries should be unaffected.

Review URL: https://chromiumcodereview.appspot.com/11458009