Add IncludeForRequestURL method to CanonicalCookie.
Move the code for checking whether to include a cookie or not from the CookieMonster to the new method in CanonicalCookie. This means the new method IncludeForRequestURL returns true when the cookie should be included for the given |url| and cookie options.
Change the CookieMonster to use the IncludeForRequestURL method of CanonicalCookie to test whether a cookie should be included or not.

TEST=CanonicalCookieTest.IncludeForRequestURL  (net_unittests)
BUG=none

Review URL: https://chromiumcodereview.appspot.com/11308272