[Sync] Display encrypted data in about:sync
For ease in debugging, we automatically decrypt encrypted specifics if we are
able when creating their string representations. This allows us to display
decrypted data in the sync node browser and data dumps.

Note that passwords don't use the normal encryption scheme, and hence are not
decrypted. This is intentional.

BUG=96225


Review URL: https://chromiumcodereview.appspot.com/11557006