Fix Javabridge crash.
We take control of the FD from used by the channel. It prevents the channel from
closing it until the channel is deleted.  https://codereview.chromium.org/11308342/
will make sure that the channel is deleted at the end and we will not have leaking
FD.

BUG=157699


Review URL: https://chromiumcodereview.appspot.com/11502002