UserPolicySigninService now checks if a user is from a hosted domain.
Added UserInfoFetcher which makes an OAuth call to the userinfo service to fetch
information about the currently signed-in user.

UserPolicySigninServices uses a UserInfoFetcher to check whether a user is
from a hosted domain before trying to fetch a DMToken, to reduce the load on
the DMServer.

BUG=141123
TBR=brettw


Review URL: https://chromiumcodereview.appspot.com/11552019