Use consistent velocity trend to start a touch scroll gesture.
In the GestureRecognizer, we previously inititated scrolling when the finger leaves a L1-norm box surrounding the position of the touch down point. As a result, the page would seem to leap by the size of the L1-norm on the first paint after entering a scrolling gesture.

This CL replaces the L1-norm box mechanism with a new mechanism for determining the start of a scroll: watch for the establishment of a consistent velocity trend across multiple touch moves. This permits determining that a scroll should start with a much smaller displacement than is needed with the L1-norm technique and eliminates the "leap" effect.

BUG=141653


Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=171934

Review URL: https://chromiumcodereview.appspot.com/11469005