Extra compile assert in ipctest.cc
Shishir ran into this earlier today, and it was confusing that this file, in an uncommon target, is the only error you get when you add a spurious semicolon to an IPC header. This compile assert is SHOUTING so that future generations do not waste so much time.

R=shishir@chromium.org,thakis@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/11568018