On Mac, use built-in pnacl files located in the build directory
The --enable-pnacl flag is supposed to enable use of "built-in" pnacl
files (as opposed to files from an extension). On windows and linux
they can just be found in the build directory, without being packaged
or shipped in an installer. However on mac, putting them in the internal
plugins directory in the bundle would also cause them to get packaged,
so for now we do a directory traversal to find them outside the
package, similarly to the way logs are handled currently.

R=jvoung@chromium.org,brettw@chromium.org
BUG= http://code.google.com/p/nativeclient/issues/detail?id=2365


Review URL: https://chromiumcodereview.appspot.com/11555037