Revert the hard-coded triggering for the Sirefef malware.
This reverts a change that caused the Sirefef public key to be revoked in the
code. That was good for testing, but the messaging is in a different patch
which is staying on the branch and, by not having the revocation built into the
code, we can switch it off and on later with a CRLSet push.

BUG=163227

TBR=agl@chromium.org
Review URL: https://codereview.chromium.org/11570037