Ensure that windows/views in Aura/Ash get created in the last active desktop on Windows.
The creation context is not fully threaded through to all places which need them. For e.g. if we instantiate the
settings widget in Chrome Ash on Windows 8, the widget ends up getting instantiated in desktop.

As a short term workaround, we use the GetActiveDesktop helper in the host_desktop.cc file which uses the active desktop
monitor to figure out the host type for the widget getting created. Added a TODO around the code paths which use this
namely the GetHostDesktopTypeForNativeWindow and GetHostDesktopTypeForNativeView functions to remove this once we thread
context through.

BUG=151718
R=cpu

TBR=sky
Review URL: https://codereview.chromium.org/11583008