Disable the script bubble for now.
Will revisit for the next milestone once we have addressed some feedback we got.

Also remove an artificial bump in location bar event count in the Script _Badge_ unit test, that was accounting for a bump due to the Script _Bubble_. Once the Script Bubble is online again I hope to have converted it to use the Observer model so as to not have to rely on these LocationBarView updates (so I don't expect the hack to make a comeback).

BUG=153137
TEST=The script bubble should not be seen.

Review URL: https://chromiumcodereview.appspot.com/11566029