Fix PPAPI time conversion code to handle epoch times.
This fixes the time conversion code to be able to correctly handle epoch times
(e.g. when passing a PP_Time of 0). This eliminates the need to special-case
for epoch time in other code. In practice the problem this fixes only
manifests itself in test code.

BUG=

Review URL: https://codereview.chromium.org/11567013