cc: Pass flag to Skia to make path clip recording cheaper
SkCanvas::clipPath returns a bool which forces the clip to be applied
right then to see if it is empty or not.  This flag just tells Skia to
use the bounds of the path itself, which should speed up record times for
path clips.

R=nduca@chromium.org
BUG=none


Review URL: https://chromiumcodereview.appspot.com/11573049