Send compositor frame IPC with metadata.
This makes CompositorFrame contain one of two payloads: DelegatedFrameData for ubercomp and GLFrameData for the Aura model.  This also adds CompositorFrameMetadata containing information useful for positioning subwindows relative to the webpage scroll position.

The message is sent when a new switch --enable-compositor-frame-message is set.  This is enabled by default on Android and also suppresses Android-specific UpdateFrameInfo messages, which will be superceded by CompositorFrameMetadata.

BUG=152337,161945

Review URL: https://chromiumcodereview.appspot.com/11316128