cc: Create a cc::ScrollbarThemePainter and an adapter class in compositor bindings.
Instead of having the scrollbar layer talk directly to the
WebScrollbarThemePainter, we make an abstract interface in cc/ that it can talk
to. In the compositor bindings, we implement this class, and forward the
function calls made into it over to the WebKit implementation.

No change in behaviour, no new tests. This is a prerequisite for
https://codereview.chromium.org/11464041/.

BUG=123444
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11609002