Browser Plugin: Simplify BrowserPluginGuestHelper
Since BrowserPlugin was first implemented, support for cross-process navigation
was dropped but BrowserPlugin*Helper classes maintain a lot of complexity that
was no longer necessary. This CL attempts to reduce some of this complexity
by simplifying BrowserPluginGuestHelper. It is only needed now to intercept
ViewHostMsg_* before arriving to RenderViewHost. It is still a separate class
because RenderViewHostObserver's lifetime is managed by RenderViewHost and
BrowserPluginGuest's lifetime is managed by WebContentsImpl.

This refactor/simplification is a prerequisite to introducing a
BrowserPluginGuestObserver class. Further changes will come in a subsequent
CL to simplify the message routing from

BrowserPluginEmbedderHelper --> BrowserPluginEmbedder --> BrowserPluginGuest

BUG=166165
Test=BrowserPluginHostTest.*

Review URL: https://codereview.chromium.org/11606005