Get rid of unnecessary call updateImeSelection
This code was added with m18 adjustmenets that weas not upstreaming and it is causing
unnecessary update selection calls that confuses some keyboards. Removing this for now.
We will have to keep an eye on any issues this may cause, but from the previous logs
it looks like we don't need these for any legitimate reason anymore.

BUG=151496


Review URL: https://chromiumcodereview.appspot.com/11607011