Use a queue instead of timestamps to sort most-recently-used clients
Use a most-recently-used queue to sort the GPU memory manager clients.
Previously we used timestamps, but this had issues with unit testing
because of timer resolution.

Maintain three lists of clients -- visible, non-visible, and non-surface
clients. The visible and non-visible lists are in most-recently-used order.
This removes the need to create a separate vector of the clients and
sort that vector at every manage call.

Store whether or not a share group has any non-hibernated clients directly
in the share group's memory tracking group, so we don't have to construct
an explicit set of the non-hibernated clients.

BUG=164947


Review URL: https://chromiumcodereview.appspot.com/11645013