Adjust fonts (part of 137349): Use standard font sizes in ui/app_list
As part of a bigger issue (crbug.com/137349) we're adjusting fonts. Changing fonts and font sizes per direction from designers and to use standard font/sizes instead of DeriveFont +/-N pixels. 

These changes affect the App List UI invoked by clicking on the App icon on the launcher. In the item view the UI is the same, we're just using a standard font name. In the search box the font size increased from 14 to 15 pixels. That was ok-d by a designer.


BUG=166491, 137349
TEST=manual: Click on App icon in launcher area; type in search text

Review URL: https://chromiumcodereview.appspot.com/11576058