Delete implementation of ChromeRenderViewObserver::DidHandleTouchEvent
The code was crashing, the implementation was wrong anyway, and the result
was unused! This is a safe fix for M25 that just deletes the whole method
implementation. Not deleting the method because Ananta already has a change
in progress that reimplements it the correct way and makes use of it.

BUG=162433
TBR=darin


Review URL: https://chromiumcodereview.appspot.com/11606011