cc: Add pending tree for impl-side painting
Do tree synchronization to pending tree.  Inform scheduler when there is a pending tree so that vsync can continue to be scheduled.

Right now everything is hooked through redraw, so canDraw is true when there is a pending tree, prepareToDraw is the one that checks for tree swap, and TileManager::DidFinishTileInitialization ends up calling setNeedsRedraw.  I'm not sure this is ideal, but it's a start.

BUG=155209

Review URL: https://chromiumcodereview.appspot.com/11569040