Allow the resourceProvider to be NULL when evicting textures. In that case, make sure that there aren't any textures.
This comes up in situations where we are re-initializing the renderer twice after a lost context. We delete all of the textures (with the old resourceProvider) after the first context lost. If renderer initialization fails, we may pass in a NULL resourceProvider to a second eviction call. In this case, all of the textures must be gone anyway.

Fix a bug in single thread proxy where we wouldn't allow repeated texture purges (that used to be unsupported, but is supported now).

BUG=


Review URL: https://chromiumcodereview.appspot.com/11614034