Discard zero-velocity fling-start events from trackpad.
When a two-finger scroll ends from the trackpad (e.g. by the user
releasing a finger), it sends a zero-velocity fling-start event. This
is necessary for the overscroll controller to behave properly. However,
since this is a zero-velocity fling, it can (and should) be dropped so
that it doesn't reach the renderer.

BUG=166943

Review URL: https://codereview.chromium.org/11649041