Disable PPAPINaClGLibcTest on Windows since they're failing on the windows bots.
Note that all these tests are failing on first run. Some of them are also failing on the gtest retry step. Tests in browser_test are supposed to be shardable. Note that this doesn't mean that the tests should move to interactive_ui_tests, since that's supposed to be minimal by definition as otherwise it'll be too slow. So the tests need to be fixed so that they run in parallel.

BUG=162094
Review URL: https://codereview.chromium.org/11637054