Ensure that all resources are torn down before the routing.
This force-deletes the resources before the various PpapiHosts are torn down. This ensures that all pointers are still valid at the time of resource destruction, allowing resources to do nontrivial things in their destructors. This part of the patch is forward-lookind and doesn't fix any current known bugs.

Also force-unbind all the graphics from the instance destructor as well, which ensures that the Graphics2D host doesn't have a pointer to a bound instance that outlasts that instance (which fixes the below-mentioned bug).

BUG=166353


Review URL: https://chromiumcodereview.appspot.com/11640030