cc: When the transform's scale is unknown, draw into render surfaces at the scale of the screen.
When the transform's scale is not known, we have to use some fixed value. We
were using 1.0 but this means if the transform is close to the identity, the
layers in the surface are rendered small, and then scaled up when the surface
is drawn to the screen by the device scale factor and page scale factor. This
makes for fuzziness.

So, when the transform's scale is not known we should use the device scale
and page scale factors as the default value. This guarantees that the transforms
from the contents of the surface's subtree to the screen will not require
scaling up more than any CSS scales specified by the HTML.

Tests:
LayerTreeHostCommonTest.verifySurfaceLayerTransformsInHighDPI

BUG=167021

Review URL: https://chromiumcodereview.appspot.com/11635037