Verify that cameras support desired formats.
Add a check whether a camera supports formats that we are interested in.
Without this it'd be possible that we'd favor a camera that doesn't support
desired formats over one that would, but with a lower video node number.

This also adds MJPEG as an option even if the resolution is lower than
minimum MJPEG resolution as the last choice, while keeping it the first
choice for resolutions larger/equal to minimum MJPEG resolution.

BUG=166299


Review URL: https://chromiumcodereview.appspot.com/11585013