PpapiThread should handle child process messages.
The current code overrides ChildThread::OnMessageReceived(), but doesn't handle child process messages.
One consequence is that we ignore the browser message to shut down ppapi processes.

With this CL, a ppapi process should be shut down when the plugin is not used by any tab for 30 seconds.

BUG=None
TEST=None


Review URL: https://chromiumcodereview.appspot.com/11633051