Fix AudioPipeReader to handle EOF properly.
PulseAudio is started on demand. When PulseAudio is not started read() alaways
returns 0 (i.e. EOF). Normally pulseaudio is started on demand by libpulse so that
it can be restarted after it's stoped or crashed. Previously AudioPipeReader::DoCapture()
didn't handle properly the case when read() returns 0, which would cause it to get into 
indefinite loop when pulseaudio is not started.

BUG=167801

Review URL: https://chromiumcodereview.appspot.com/11744009