Simplify ExecuteJavaScript* functions.
Remove the "Java" and rename to ExecuteScript*.  This better matches
conventions elsewhere in the codebase.

Introduce ExecuteScriptInFrame* variants for the less common case
where script execution in a subframe is desired.

Make it possible to pass the ExecuteScript* family of functions either
a WebContents pointer or a RenderViewHost pointer to further simplify
callsites.

R=jam@chromium.org
BUG=none


Review URL: https://chromiumcodereview.appspot.com/11753009