Removing the JumpRelative, Call, CallRelative and Return commands.
This removes the extra commands from the command buffer. I've left the
Jump struct and handler in place as it's used to advance the buffer
if there isn't enough room at the end to place the next command.

TESTING=Ran the gpu_unittests and everything worked correctly.


Review URL: https://chromiumcodereview.appspot.com/11613021