Initial UniquePositions implementation
Introduces a new class that represents an item's position within an
arbitrarily ordered list.  It's currently biased towards bookmarks, but
it could be made more generic if there was a need for it.

This class supports inserting before, after, or between other
UniquePosition items.  It never runs out of position values, though it
can use up unbounded space in the worst case.

In addition to the basics mentioned above, it also has support for other
functions that are particularly useful for bookmark sync.  The positions
are serializable and de-serializable, for storage and transimission.
They also support conversion to and from int64 values, allowing them to
operate with old-style positions.

The class is not currently used anywhere.  This commit includes only its
implementation and some tests.

BUG=145412


Review URL: https://chromiumcodereview.appspot.com/11569045