Fix hang in Android WebView release build tests
SurfaceTexture.detachFromGLContext is not present in ICS and the old
workaround of using a NullCompositor lead to the test creating
TextureImageTransportSurface in ImageTransportSurface::CreateSurface.
This path is wrong and not well tested in single-process mode and recently
lead to a deadlock hang in release build only tests (see bug).

Fixed by creating the real compositor but just avoid calling the missing
method in ICS. The Java method is loaded lazily so this does not cause
an error at run time.


BUG=167236,161864
Android only change. Ran through Android trybots.
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11787031