Implemented BluetoothExtensionFunction.
From discussion with Bryan (@bryeung) and Scott (@keybuk), we decided to make
the bluetooth extension functions asynchronous because it takes time to
initialize BluetoothAdapter. So I implemented BluetoothExtensionFunction, which
is the base class for all the subclasses and it calls DoWork() when the
bluetooth adapter is ready to use.

After this CL is submitted, I will turn
BluetoothAdapterFactory::DefaultAdapter() to receive a callback and run the
callback when the bluetooth adapter is ready.

TBR=sky@chromium.org
BUG=135470


Review URL: https://chromiumcodereview.appspot.com/11743024