sync: Start moving away from PREV_ID and NEXT_ID
This change replaces serveral instances of Get(PREV_ID) and Get(NEXT_ID)
with calls to the newly introduced GetPredecessorId() and
GetSuccessorId().

This is done in anticipation of the change that will remove PREV_ID and
NEXT_ID entirely.  By making this change in advance, that patch will be
smaller and easier to review.

BUG=145412,126505


Review URL: https://chromiumcodereview.appspot.com/11637053