Connect to DesktopEnvironment's stubs only when audio/video schedulers are about to be created.
This CL localizes usage of DesktopEnvironment to a single method that creates audio/video schedulers and hook up the event executor. This prepares further refactoring of DesktopEnvironment that should become a factory for classes interacting with the specific desktop environment.

Related changes in this CL:
  - MouseClampingFilter receives screen size on the proper thread.
  - ClientSessionTest.* tests now run asynchronously to support MouseClampingFilter change.
  - ClientSession does not expose DesktopEnvironment.
  - InputEventTracker can be re-attached to a different InputStub.

BUG=104544


Review URL: https://chromiumcodereview.appspot.com/11781003