Refactor SetOmahaExperimentLabel out of gcpai and into install_util.
Change the method to no longer write to both App GUIDs, just the one that matches the distribution of the caller. This means that GCAPI will only write to the Google Chrome product's client state key, and not to the Chrome Binaries.

Replace SetOmahaExperimentLabel in the GCAPI code with a more specific version, and have SetOmahaExperimentLabel perform the general clobber-write to the registry.

This is in preparation for future work in using the experiment_labels to transmit Chrome Variations.

BUG=160251


Review URL: https://chromiumcodereview.appspot.com/11280067