Make it so the windows store protocol is properly handled
The core problem is that the key that controls the execution of a protocol
HKEY_CLASSES_ROOT\<protocol>\Shell\Open\Command

Has changed, we expect the Default value to be a valid command like foo.exe %1
but this is no longer the case. This value can be empty and other values will
control what happens, for instance DelegateExecute.

This CL also whitelists the protocol, alternatively we could hardcode a meaniful
string to present to the user but that presents localization issues. We take the
same approach as IE and just launch the windows store if we see this protocol.

BUG=159096
TEST=see bug
Review URL: https://codereview.chromium.org/11316038