Notify policy change observers via a task
This prevents unexpected reentrancy in OnUpdatePolicy() in the case that an observer changes policy (leading to a recursive call to OnUpdatePolicy()).

This avoids a crash that happens if we change the SyncDisabled policy, since
this signs out the user and modifies policy while we are still iterating through
the PolicyBundles in MergeAndTriggerUpdates().

BUG=155486


Review URL: https://chromiumcodereview.appspot.com/11667006