cc: Add RasterWorkerPool class.
This moves thread pool related code out of the tile manager and into a
new RasterWorkerPool class. This doesn't change the behavior of the tile
manager in any way it just separates tile management logic from the
thread handling logic, which results in a significant cleanup of the
tile manager.
    
BUG=155209
TEST=manual

Review URL: https://chromiumcodereview.appspot.com/11593030