sync: speculative fix for bug 165561
The SyncManager will stop the scheduler as part of StopSyncingForShutdown, which
happens well before things like the invalidator are torn down / told to stop
interacting with the scheduler.  As such, we can wind up scheduling nudges on a lame
duck scheduler.  Some of the stacks associated with 165561 clearly show browser
shutdown happening on the UI thread and OnIncomingInvalidation happening on the
sync thread, which will crash on a second invocation as pending_nudge will be
set to a job that is dropped by ScheduleSyncSessionJob.

BUG=165561


Review URL: https://chromiumcodereview.appspot.com/11827013