overscroll: Make sure screenshot is taken correctly for swapped out RenderViews.
When a RenderView is swapped out, it navigates to a swapped-out url
(swappedout://), and thus corrupts the paint content of the actual view. To fix
this issue, prevent a swapped-out RenderView from painting. Also, make sure the
screenshot is taken in the browser side before the RenderViewHost is swapped
out.

BUG=160668

Review URL: https://codereview.chromium.org/11776010