Delay transient downloads a few seconds before displaying.
Downloads that are temporary or will be opened automatically will be
removed from the shelf when they complete. If these are quick
downloads, the user will not be able to interact with them on the
shelf. Delay showing these on the shelf by a few seconds to avoid
showing quick downloads on the shelf.

BUG=164710


Review URL: https://chromiumcodereview.appspot.com/11689003