Duplicate setup.exe into an identical executable and point Active Setup to that executable instead.
Active Setup is a Windows "Run-once-for-each-user-at-next-login" mechanism which we use to install per-user shortcuts on system-level installs.

It currently points to setup.exe --configure-user-settings --...

Windows' clever heuristics detect that this is an installer solely based on the process name and tries to elevate (this is mitigated on Vista since we set requestedExecutionLevel="asInvoker" in setup's manifest; on XP however this results in a blocking "Install Program as Other User" dialog at log in...).

BUG=166473
TEST=
On Windows XP:
0) Have User A (admin) and User B (non-admin, logged out)
1) Install system-level Chrome from User A
2) Log in to User B
3) See that the per-user Quick Launch shortcut has been installed (and that you weren't prompted by the "Install Program as Other User" dialog).

More Active Setup testing can be done as per the TEST block on https://codereview.chromium.org/11465025/.

Review URL: https://chromiumcodereview.appspot.com/11733007