Convert PendingQueue.pending_commits from a list to a dict.
It will require a wipe of the CQ's state.

This CL requires a fair number of list copies, which is not super efficient,
but it ensures that the pending_commits dict is always in a consistent state,
which is _more_ important. Using index-based entry removal in a list requires
it to be coordinated, which wasn't 100% the case.

It keeps PendingQueue.pending_commits a "public member" so it is properly
serialized but the member functions should be used to access it to ensure
correct type and ordering.

Updated the test data with:
import json
for f in ('chromium_state.json', 'chromium.11299256.json'):
  a = json.load(open(f))
  a['pending_commits'] = dict((str(i['issue']), i) for i in a['pending_commits'])
  open(f, 'w').write(''.join(l.rstrip() + '\n' for l in json.dumps(a, indent=2, sort_keys=True).splitlines()))

R=csharp@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/11821054