Allow dragging a download only after the download is complete.
This was already the case on views and cocoa. However GTK was allowing a drag
after the large icon finishes loading. This CL also updates callers to
use GetTargetFilePath() on the DownloadItem instead of GetFullPath().
For a completed download, both methods are equivalent, but we prefer
the former as it is clear that the path being requested is the final
target path.

BUG=116551


Review URL: https://chromiumcodereview.appspot.com/11801022