Add test to ensure Chrome welcome page loads on Android.
There are no tests currently that ensure that any chrome:// pages load
for Android, and there have been times where they have stopped loading,
but no test caught this.

This adds a test that loads chrome://welcome/ in the Chromium test
shell, and ensure it does not show an interstitial page, and that the
title and URL look sane.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/11823063