Add a new code/message path for moving an insertion handle
On Android, we've been using selectRange for moving an insertion handle.
However, we don't really want the behavior of selectRange (for example,
dragging an insertion handle outside of a text box should not actual do
a "selectRange" on the position outside of the text box). The behavior
that we want has been added to WebKit as
WebFrame::moveCaretSelectionTowardsWindowPoint.

This change introduces the code/message path from the Browser side
InsertionHandleController to WebKit similar to the current SelectRange
path.

BUG=165244,163979,165661


Review URL: https://chromiumcodereview.appspot.com/11593011