Duplicate handles so each SharedMemory object will have its own handle.
SharedMemory assumes that it owns the handle of the shared section and closes it upon destruction. If two SharedMemory objects use the same handle value they both will try to close the handle and the 2nd attempt will fail. Usually this error is ignored but Application Verifier detects it easily.

BUG=169030


Review URL: https://chromiumcodereview.appspot.com/11833004