sync: Represent UniquePositions as a protobuf
Although this has little practical effect right now, it gives us the
flexibility to convert to more complex protobuf representations in the
future.

The reuse of an existing protobuf field would normally be a very bad
idea, but we're fairly certain that no one has actually made use of the
old definition and that we can clean up all references to it.

BUG=145412


Review URL: https://chromiumcodereview.appspot.com/11825017