Fix LevelDB histogram code.
The old code misused the macros provided by histogram.h such
that a DCHECK failed on the bots. UMA_HISTOGRAM_ENUMERATION
et al use a static variable to retrieve the Histogram
objects, so a single instance of that macro can't be used to
retrieve multiple histograms. The old code tried to retrieve
both LevelDBEnv.IOError and LevelDBEnv.IDB.IOError. This CL
forgoes use of the macro and keeps a pointer to the
Histogram objects around.

http://wkb.ug/106135 will have to reland to make use of this.

BUG=164755


Review URL: https://chromiumcodereview.appspot.com/11860015