app_list: Fix UpdateIcon crash for app update.
When an app is unloaded for updating, UpdateIcon() is called. However,
GetExtension() returns NULL at this time and triggers a crash in HasOverlay().
Since HasOverlay() info is not changing, cache it in a |has_overlay_| to fix
the crash.

Also fix a potential stale title/icon problem when extension app is updated.

BUG=157996,169463
TEST=Verify app list should not cause a crash when an app is updated.

R=sky@chromium.org,benwells@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11885006