Decouple texture upload checks, tree activation, and drawing.
This prevents us from drawing more often than we need to
when impl-side-painting is enabled.

It removes the concept of scheduling checkForCompletedSetPixels, instead
checking for completed uploads 1) before determining whether or not to make
the pending tree active and 2) before drawing.

Tree activation is handled in two places now: 1) before calculating
animations and 2) before a vsync tick.

When a pending tree is activated, a request to draw is made.

BUG=168724


Review URL: https://chromiumcodereview.appspot.com/11823043