Encrypted Media: Enforcing the CDM to decode audio into S16 integers.
After the new FFmpeg roll, FFmpeg decodes into float instead of integers. Therefore, it may be confusing about what the CDMs are returning. In this CL, we limit the DecryptingAudioDecoder to support S16 only to work around this issue. All CDM implementations are expected to return S16.

This is a short term fix for the audio corruption in CDMs. In the future, we need to be able to accept any format returned by the CDMs (see http://crbug.com/169105).

BUG=168862
TEST=media_unittests pass; clearkeycdm audio works well.


Review URL: https://chromiumcodereview.appspot.com/11778079