Renderer form_autofill_util changes to support Checkboxes and Radiobuttons.
Checks for checkable elements while reading the loaded page.

This is attempt no. 3 to submit this change in. Attempt no. 2 failed as it regressed on win7-* builds, while it was good on linux-release builds. This CL is one of the smaller CLs created out of r176189 in an attempt to narrow down the code effecting win7-* builds.


BUG=157636


Review URL: https://chromiumcodereview.appspot.com/11875019