Make nacl_integration find the newest chrome binary.
Previously, it would take the first one that is found from a handmade
preference list.  This could mean that we end up testing a chrome
binary that is left over from a non-ninja build even if the
compile was done with ninja, just because the preference list
has non-ninja paths earlier than the ninja paths.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=3257


Review URL: https://chromiumcodereview.appspot.com/11884024