Eliminate media::Buffer as a base class for media::DecoderBuffer and media::DataBuffer.
It was never a good idea in the first place.

Our usage is exclusively with DecoderBuffers or DataBuffers. There's never a case where we benefit from using Buffer as a base class aside from hiding GetWriteableData(), however it's not a compelling enough reason to keep Buffer around.

BUG=169614
TBR=dmichael

Review URL: https://codereview.chromium.org/11880008