CrOS: Revert change to OOM priority adjustment interval
When switching tabs we adjust the OOM priority of the focused tab's renderer
so that the focused tab is protected from the next discard. We do this after
a delay to avoid slowing down tab-switch time. I incorrectly increased this
delay yesterday thinking it meant something else.

BUG=none
TEST=none

Review URL: https://codereview.chromium.org/11933004