Send the ClientHello handshake message.
Add most of the fields to ClientHello. Add the
QuicCryptoClientStream::CryptoConnect method and have
QuicClientSession::CryptoConnect call it. Add a unit test for
QuicCryptoClientStream to check a ClientHello message by parsing
it and checking the fields.

Define the value in CryptoTagValueMap as string instead of
StringPiece. This causes CryptoTagValueMap to copy the values,
which is less efficient but much safer.

Merge internal CLs: 39503243, 39692710

R=rch@chromium.org
BUG=none
TEST=net_unittests

Review URL: https://chromiumcodereview.appspot.com/11633030