Fix for a crasher when starting a renderer.
A crash dump from the field reports that the CrashDumpManager in
ChromeContentBrowserClient is not initialized when
GetAdditionalMappedFilesForChildProcess is called.
This is because the CrashDumpManager is initialized when a RenderProcessHost
is created, but it could be the case that a BrowserChildProcessHost gets
created first and calls GetAdditionalMappedFilesForChildProcess before
any RenderProcessHost was created.
Making the CrashDumpManager a singleton initialized early on to prevent this
crasher.

BUG=163456
TEST=None


Review URL: https://chromiumcodereview.appspot.com/11884042