Bring back the file dialogs for chrome AURA.
This CL copies the ui\base\dialogs sources to a new directory ui\shell_folders. This is
done due to problems with gcl not working correctly in conjunction with svn mv and local edits leading
to tryjob issues.

The code in ui\shell_folders is not compiled at the moment.

This will be compiled when CL https://codereview.chromium.org/11878031 lands.

BUG=170483
TBR=ben