cc: Add point-based UV coordinate on TextureLayer
Previously, the uv coordinates was using gfx::RectF as a struct.
This was limitating because it could not handle negative width
and heights. Soring them as 2 gfx::PointF for the Top-Left and
Bottom-Right points aleviate that restriction.
The current implementation of the phone UI rely on this feature.

BUG=168927

Review URL: https://chromiumcodereview.appspot.com/11783094