Browser Plugin: Simplified Embedder Focus state tracking
BrowserPlugin should be grabbing the embedder's focus state straight from the render_view if it exists instead of caching the focus state in a member variable. This simplifies code a bit and is a prerequisite to the new way of allocating instance IDs and initializing state coming in a subsequent patch.

BUG=140316

Review URL: https://chromiumcodereview.appspot.com/11959016