[Media Gallery] Remove extension gallery watch count details from GalleryFilePathWatcher.
An extension can call addGalleryWatch() multiple times (in the same or different extension instance) to watch the same gallery. Previously, GalleryFilePathWatcher maintained a watch count to track the references. This reference count is not required because the extension instances share the same preference settings. If the extension instance A revoked a gallery permission, it gets reflected in extension instance B.

If multiple instances of the same extension are launched, irrespective of the extension instance that added the gallery watch, when the watched gallery changes, all the instances of the extension should be notified about the gallery changed event.

If an extension instance calls addGalleryWatch() on an already watched gallery, the request will be treated as no-op and a success result is set on the API response.

BUG=144491
TEST=none

Review URL: https://chromiumcodereview.appspot.com/11966003