Merge 176908 - Fix a problem where OpenDevice would open the wrong device if multiple are avaiable.
The issue comes from the fact the OpenDevice should not select the device
based on the user prefence but on the client request passed in the
requested_device_id. Only if this is not possible will it use the
default device.

BUG=166925
TEST=Manual: Plug multiple capture devices and try to select each one in flash settings.


Review URL: https://chromiumcodereview.appspot.com/11820050

TBR=pastarmovj@chromium.org
Review URL: https://codereview.chromium.org/11975042