GestureFlingCancelsFiltered unit test for touchpad
After introducing source device for gesture events,
RenderWidgetHostTest.GestureFlingCancelsFiltered unit test is only run for
touchscreen events. The test is now changed to run for both touchpad and
touchscreen, so that if in the future touchpad and touchscreen events are
treated differently, the tests cover both cases.

BUG=169122

Review URL: https://chromiumcodereview.appspot.com/11827037