[NaCl SDK] Build a naclports bundle (linux only)
Add new build steps to build_sdk that pull down and build
selected naclports from the naclports repo and include them
in a naclports bundle.

This is the initial version which doesn't yet support Debug
builds and has a very small list of ports that it includes.

BUG=168693

Review URL: https://chromiumcodereview.appspot.com/11952002