Add ui tests to android webkit-latest bot
This bot is supposed to be a canary for the main waterfall,
but it's configuration got out of sync.  The android testers
on main waterfall are running a variety of android
instrumentation tests, but this bot was only running unit
tests.  As a result, the latest webkit deps roll broke the
android testers on main waterfall.

Make this bot a 'copied' configuration to keep the set of tests
in sync from now on.

TBR=yfriedman

Review URL: https://codereview.chromium.org/11938011