Show app launcher after install an app, if the app launcher is installed.
THe app launcher was previously shown when the launcher was enabled via a flag, but not when it was enabled by being installed.

This change introduces a platform independent way of determining if the app launcher is enabled. This will also be used in the New Tab Page.

TBR=sky@chromium.org for gypi change
BUG=152854


Review URL: https://chromiumcodereview.appspot.com/11975035